requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
PyJWT==2.9.0
pytz==2024.1
sqlparse==0.5.1
whitenoise==6.7.0
